feat: add `single-runner` input to release.yaml
This patch introduces a input parameter to the release action that when
enabled will disable the multiple runner release job and allow the
entire process to be ran on x86 runners.

This will be painfully slow for most releases, but since arm runners
aren't available for private repositories, we will need this to do some
internal component releases from the one private repository we have.